# contracts/control-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: AstrLink Local Control API
  version: 1.0.0
  summary: Stable local contract between the AstrLink desktop application and Core.
  description: |
    This API is a local control plane; it is not part of the public inference API.
    Core binds the inference plane to `127.0.0.1:18317` by default and binds this
    control plane to the ephemeral address `127.0.0.1:0`. On startup Core writes
    exactly one ready JSON line to stdout containing the selected `control_url`.
    Launchers must use that value rather than guessing a port.

    The bootstrap health, version, and capability operations are intentionally
    unauthenticated and disclose no configuration. Every operation that reads or
    mutates user configuration, events, or request records requires the per-start
    local control token. The token must be delivered through an inherited pipe or
    equivalent protected IPC; it must not be placed in process arguments or the
    environment.

    Authenticated callers hold one of two roles, recorded on each operation as
    `x-astrlink-role`. The desktop's per-start control token is `operator` and may
    call everything. Agent tools reach Core through the same-user control socket or,
    on Windows, a separate per-start observer token; both are `observer`, which may
    call only GET operations marked `observer`. Every non-GET operation, token
    reveal, and OAuth authorization session is `operator`, and an observer calling
...
    operations.

...

...
servers:
...
security:
  - controlToken: []
...
paths:
...
  /control/v1/policies/{policy_id}:
    parameters:
      - $ref: "#/components/parameters/PolicyId"
    get:
      x-astrlink-role: observer
      x-astrlink-implementation-status: implemented
      tags: [Policies]
      operationId: getPolicy
      summary: Read one policy.
      description: |
        Operators receive the full `Policy` and its ETag. Observers receive
        `PolicySummary`, which reports allowlist entries and custom regex rules
        as counts and kinds only, and no ETag: it hashes the full document.
      responses:
        "200":
          description: Policy details.
          headers:
            ETag:
              description: Sent to operators only.
              schema:
                type: string
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Policy"
                  - $ref: "#/components/schemas/PolicySummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      x-astrlink-role: operator
      x-astrlink-implementation-status: implemented
      tags: [Policies]
      operationId: updatePolicy
      summary: Update a policy.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/PolicyPatch"
      responses:
        "200":
          description: Updated policy.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      x-astrlink-role: operator
      x-astrlink-implementation-status: planned
      tags: [Policies]
      operationId: deletePolicy
      summary: Delete a policy.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Policy deleted. Existing audit records are not implicitly deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
  /control/v1/privacy-models:
    get:
      x-astrlink-role: observer
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: listPrivacyModelInstallations
      summary: List persistent model-and-variant installations and current progress.
      responses:
        "200":
          description: Installed, downloading, paused, or failed model variants.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      x-astrlink-role: operator
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: installPrivacyModel
      summary: Start or retry one immutable model-and-variant installation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyModelInstallRequest"
      responses:
        "202":
          description: Download accepted.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
  /control/v1/privacy-models/{installation_id}:
    parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/PrivacyModelInstallationId"
    get:
      x-astrlink-role: observer
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: getPrivacyModelInstallation
      summary: Read one installation and its current progress.
      responses:
        "200":
          description: Installation status.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      x-astrlink-role: operator
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: deletePrivacyModelInstallation
      summary: Cancel a download or remove one installed model variant.
      description: Deletion is rejected while the privacy policy references the installation.
      responses:
        "204":
          description: Download cancelled or installation removed.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /control/v1/privacy-models/{installation_id}/pause:
    parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/PrivacyModelInstallationId"
    post:
      x-astrlink-role: operator
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: pausePrivacyModelInstallation
      summary: Pause a model download and retain its downloaded files.
      responses:
        "200":
          description: Current installation status; a completed download remains ready.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
  /control/v1/privacy-models/{installation_id}/resume:
    parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/PrivacyModelInstallationId"
    post:
      x-astrlink-role: operator
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: resumePrivacyModelInstallation
      summary: Resume a paused or failed model download from retained files.
      responses:
        "200":
          description: Current installation status; a completed download remains ready.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  securitySchemes:
    controlToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-control-token
      description: Per-start local token delivered over protected IPC; never persist it in configuration or logs.
  parameters:
...
    PolicyId:
      name: policy_id
      in: path
      required: true
      schema:
        $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag from the latest GET; prevents lost updates.
      schema:
        type: string
        minLength: 3
        maxLength: 128
  headers:
    NoStore:
      description: Sensitive local-control responses must not be cached.
      schema:
        type: string
        examples: [no-store]
...
    ETag:
      description: Strong entity tag for the current resource representation.
      schema:
        type: string
  responses:
...
    BadRequest:
      description: Malformed request, query, or cursor.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid local control token.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    PreconditionFailed:
      description: If-Match does not match the current resource representation.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: The request is well-formed but violates a contract or safety invariant.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Policy:
      type: object
      required:
        - id
        - name
        - enabled
        - priority
        - detector
        - local_model_id
        - min_confidence
        - regex_source
        - custom_regex_rules
        - kind_rules
        - allowlist_rules
        - match
        - request_action
        - response_action
        - response_restore
        - restore_tool_arguments
        - placeholder_notice
      properties:
        id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        name:
          type: string
          minLength: 1
          maxLength: 128
        enabled:
          type: boolean
        priority:
          type: integer
          minimum: 0
          maximum: 1000000
        detector:
          $ref: "#/components/schemas/PolicyDetector"
        local_model_id:
          oneOf:
            - $ref: "#/components/schemas/PrivacyModelInstallationId"
            - type: "null"
          description: Required to reference a ready installation when detector is local_model; null for regex.
        min_confidence:
          type: number
          minimum: 0
          maximum: 1
          default: 0.8
          description: >
            Inclusive confidence threshold for local-model findings. Candidates
            below this value are suppressed before the request action is chosen.
            Regex findings do not use this threshold.
        match:
          $ref: "#/components/schemas/PolicyMatch"
        request_action:
          $ref: "#/components/schemas/PolicyAction"
        response_action:
          type: string
          const: allow
          readOnly: true
        response_restore:
          type: boolean
          default: true
          description: >
            When request_action is redact, restore this request's placeholders in
            the upstream response before returning to the client. Default true.
        regex_source:
          $ref: "#/components/schemas/PolicyRegexSource"
        custom_regex_rules:
          type: array
          maxItems: 64
          default: []
          items:
            $ref: "#/components/schemas/PolicyRegexRule"
          description: >
            Persisted even when regex_source is builtin so switching modes does
            not discard edits. Applied only when detector is regex and
            regex_source is custom; replaces the built-in Regex catalog.
        kind_rules:
          type: array
          items:
            $ref: "#/components/schemas/PolicyKindRule"
          description: >
            Normalized by Core to exactly one entry per PrivacyKind, so a
            partial patch never leaves the effective shape of a kind ambiguous.
        allowlist_rules:
          type: array
          maxItems: 128
          default: []
          items:
            $ref: "#/components/schemas/PolicyAllowlistRule"
          description: >
            Matched values are exempt from redaction and reach the upstream and
            the local agent unchanged.
        restore_tool_arguments:
          type: boolean
          default: true
          description: >
            Restore placeholders inside tool and function call arguments as well
            as visible text. The tool call runs in the operator's own agent
            harness, so it needs the real value; leaving this off while any kind
            uses the natural style lets a plausible-looking stand-in be written
            to disk or fetched over the network as if it were genuine.
        placeholder_notice:
          type: boolean
          default: true
          description: >
            When a request emits at least one token-shaped placeholder, prepend a
            short note to the upstream system instruction explaining that the
            markers are opaque and must be copied verbatim. The note never
            carries the mapping. Requests that emit only natural stand-ins skip
            it, since those need no explanation.
      additionalProperties: false
    PolicySummary:
      type: object
      description: >
        The observer view of a policy. Allowlist entries and custom regex
        patterns can name the values the policy guards, so they are reported as
        counts and kinds only.
      required:
        - id
        - name
        - enabled
        - priority
        - detector
        - local_model_id
        - min_confidence
        - regex_source
        - custom_regex_rules
        - enabled_kinds
        - allowlist
        - match
        - request_action
        - response_action
        - response_restore
        - restore_tool_arguments
        - placeholder_notice
        - skip_tool_declarations
        - inspect_additional_tools
      properties:
        id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        name:
          type: string
          minLength: 1
          maxLength: 128
        enabled:
          type: boolean
        priority:
          type: integer
          minimum: 0
          maximum: 1000000
        detector:
          $ref: "#/components/schemas/PolicyDetector"
        local_model_id:
          oneOf:
            - $ref: "#/components/schemas/PrivacyModelInstallationId"
            - type: "null"
        min_confidence:
          type: number
          minimum: 0
          maximum: 1
        regex_source:
          $ref: "#/components/schemas/PolicyRegexSource"
        custom_regex_rules:
          type: object
          required: [count, kinds]
          properties:
            count:
              type: integer
              minimum: 0
            kinds:
              type: array
              uniqueItems: true
              items:
                $ref: "#/components/schemas/PrivacyRegexDetectorKind"
          additionalProperties: false
        enabled_kinds:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/PrivacyKind"
        allowlist:
          type: object
          required: [count, by_type]
          properties:
            count:
              type: integer
              minimum: 0
            by_type:
              type: object
              propertyNames:
                enum: [literal, domain_suffix, cidr]
              additionalProperties:
                type: integer
                minimum: 1
          additionalProperties: false
        match:
          $ref: "#/components/schemas/PolicyMatch"
        request_action:
          $ref: "#/components/schemas/PolicyAction"
        response_action:
          type: string
          const: allow
        response_restore:
          type: boolean
        restore_tool_arguments:
          type: boolean
        placeholder_notice:
          type: boolean
        skip_tool_declarations:
          type: boolean
        inspect_additional_tools:
          type: boolean
      additionalProperties: false
...
    PolicyPatch:
      type: object
      minProperties: 1
      properties:
        enabled:
          type: boolean
        detector:
          $ref: "#/components/schemas/PolicyDetector"
        local_model_id:
          oneOf:
            - $ref: "#/components/schemas/PrivacyModelInstallationId"
            - type: "null"
        min_confidence:
          type: number
          minimum: 0
          maximum: 1
        regex_source:
          $ref: "#/components/schemas/PolicyRegexSource"
        custom_regex_rules:
          type: array
          maxItems: 64
          items:
            $ref: "#/components/schemas/PolicyRegexRule"
        kind_rules:
          type: array
          items:
            $ref: "#/components/schemas/PolicyKindRule"
        allowlist_rules:
          type: array
          maxItems: 128
          items:
            $ref: "#/components/schemas/PolicyAllowlistRule"
        request_action:
          $ref: "#/components/schemas/PolicyAction"
        response_restore:
          type: boolean
        restore_tool_arguments:
          type: boolean
        placeholder_notice:
          type: boolean
      additionalProperties: false
...
    PrivacyModelInstallRequest:
      type: object
      required: [repo_id, revision, variant_id, label_mapping]
      properties:
        repo_id:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,95}/[A-Za-z0-9][A-Za-z0-9._-]{0,95}$"
        revision:
          type: string
          pattern: "^[0-9a-f]{40}$"
        variant_id:
          type: string
          pattern: "^[a-z][a-z0-9_]{1,63}$"
        label_mapping:
          $ref: "#/components/schemas/PrivacyModelLabelMapping"
      additionalProperties: false
    PrivacyModelInstallation:
      type: object
      required:
        - id
        - source
        - catalog_id
        - catalog_source
        - name
        - license
        - languages
        - repo_id
        - revision
        - variant_id
        - variant_name
        - quantization
        - adapter
        - status
        - bytes_downloaded
        - bytes_total
        - estimated_ram_bytes
        - error
        - label_mapping
        - installed_at
      properties:
        id:
          $ref: "#/components/schemas/PrivacyModelInstallationId"
        source:
          type: string
          enum: [catalog, custom, local]
        catalog_id:
          oneOf:
            - $ref: "#/components/schemas/PrivacyModelCatalogId"
            - type: "null"
        catalog_source:
          oneOf:
            - type: string
              enum: [official, community]
            - type: "null"
        name:
          type: string
          minLength: 1
          maxLength: 128
          description: Trimmed model display name; control characters are not allowed.
        license:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 64
            - type: "null"
        languages:
          type: array
          maxItems: 32
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 64
        repo_id:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,95}/[A-Za-z0-9][A-Za-z0-9._-]{0,95}$"
        revision:
          type: string
          pattern: "^[0-9a-f]{40}$"
        variant_id:
          type: string
          pattern: "^[a-z][a-z0-9_]{1,63}$"
        variant_name:
          type: string
          minLength: 1
          maxLength: 64
          description: Trimmed variant display name; control characters are not allowed.
        quantization:
          type: string
          minLength: 1
          maxLength: 32
          description: Trimmed quantization identifier; control characters are not allowed.
        adapter:
          $ref: "#/components/schemas/PrivacyModelAdapter"
        status:
          type: string
          enum: [downloading, paused, ready, error]
        bytes_downloaded:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        bytes_total:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        estimated_ram_bytes:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        error:
          oneOf:
            - type: string
              enum: [download_failed, integrity_failed, incompatible_model]
            - type: "null"
        label_mapping:
          $ref: "#/components/schemas/PrivacyModelLabelMapping"
        installed_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
      additionalProperties: false
    PrivacyModelInstallationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/PrivacyModelInstallation"
      additionalProperties: false
...